Apply PR feedbacks
